### 2022-01-03 #495 @peterdesmet
frictionless: Read and Write Frictionless Data Packages
Submitting Author Name: Peter Desmet
Submitting Author Github Handle: @peterdesmet
Other Package Authors Github handles: @damianooldoni
Repository: https://github.com/frictionlessdata/frictionless-r
Version submitted: 0.9.0
Submission type: Standard

Version accepted: TBD

---

```
Package: frictionless
Title: Read and Write Frictionless Data Packages
Version: 0.9.0.9000
    person("Peter", "Desmet", , "peter.desmet@inbo.be", role = c("aut", "cre"),
    person("Damiano", "Oldoni", , "damiano.oldoni@inbo.be", role = "aut",
    person("Research Institute for Nature and Forest (INBO)", , ,
           "info@inbo.be", role = c("cph"))
  )
Description: Read and write Frictionless Data Packages. A Data Package
  (<https://specs.frictionlessdata.io/data-package/>) is a simple container
  format and standard to describe and package a collection of (tabular) data.
  It is typically used to publish FAIR and open datasets.
    https://frictionlessdata.github.io/frictionless-r/
    assertthat,
    glue,
    httr,
    readr (>= 2.1.0),
    stringr
    hms,
```

	- [x] data validation and testing (listed as [category](https://devguide.ropensci.org/policies.html#package-categories), but not in issue template)

`frictionless` allows users to read and write [Frictionless Data Packages](https://specs.frictionlessdata.io/data-package/), an open and general-purpose standard to structure and describe (tabular) datasets, typically used to publish FAIR datasets. The package allows users to read (local and remote) Data Packages (**data retrieval**), load its data resources in data frames (**data extraction**), return errors if the Data Package is malformed (**data validation and testing**), add data frames as new resources (**data munging**) and write Data Packages back to disk (**Data deposition**).

Anyone who wants to read or create datasets structured as Frictionless Data Packages. The community is referred to as the [Frictionless Data community](https://frictionlessdata.io/) and typical includes researchers, data scientists and data engineers, often interested in (publishing) open data.

Yes, [datapackage.r](https://cran.r-project.org/web/packages/datapackage.r/index.html): it has an object-oriented design (using a `Package` class) and offers validation. `frictionless` on the other hand allows users to quickly read and write Data Package data to and from R data frames, getting out of your way for the rest of your analysis. It is designed to be lightweight, follows [tidyverse](https://www.tidyverse.org/) principles and supports piping. The main functionality (reading data into data frame, adding a data frame as a resource to a package, writing a Data Package to disk) is offered as functions, rather than the class properties in `datapackage.r`.

Note that the link to guide for authors above (in the issue template) returns a 404. It should be https://devguide.ropensci.org/authors-guide.html. I tried to use `pkgcheck` but I got `package ‘pkgcheck’ is not available for this version of R`

Note that this package falls under the [Frictionless Data Code of Conduct](https://frictionlessdata.io/work-with-us/code-of-conduct).
