### 2022-01-03 #495 @stefaniebutland
@lwinfree just fyi, here's the (start of the) rOpenSci software peer review thread for the "frictionless" R package.
Folks, Lilly Winfree is Product Manager @ Frictionless Data.
