### 2022-01-04 #495 @peterdesmet
A `codemeta.json` file has now been added.
