### 2022-01-06 3ce1f63
Use bootstrap 5 (enables search) with dark navbar
