### 2022-01-07 42b63bf
Rename list_clean to clean_list (verb first)
