+++ R/create_schema.R
@@ -116,7 +116,7 @@ create_schema <- function(df) {
   )
 
   # Remove elements that are NULL or empty list
-  schema <- list_clean(
+  schema <- clean_list(
     schema,
     function(x) is.null(x) | length(x) == 0L,
     recursive = TRUE
