### 2022-01-07 73f87f2
Rename parameter "df" to "data in check_schema()
And refer to it as such in error messages: data is more understandable to user, as often the passing of `df` is something internal.
