### 2022-01-07 b387a0d
Rename parameter "df" to "data" in create_schema()
