### 2022-01-07 c366548
Rename parameter "df" to "data" in add_resource()
