### 2022-01-07 d25c99b
Support directly adding a csv file in add_resource()
