### 2022-01-07 304bd50
Fix typo in if statement
