### 2022-01-07 89df680
Explain write behaviour clearer
