### 2022-01-07 637c4b7
Explain properties of `data`
