### 2022-01-07 0cdd6b5
Create shorter example
