### 2022-01-07 5a715b7
Allow added resource to have unsafe path
