### 2022-01-07 123fec5
Add example to vignette to add_resource() from csv files
