### 2022-01-07 a671ae6
Allow to pass arguments when reading path in add_resource()
