### 2022-01-08 4f3be71
Fix formatting bug in documentation
