### 2022-01-10 0aafa32
Mention that paths can be character vector
