### 2022-01-10 c0de649
Add examples for path vector and provided schema
