### 2022-01-10 58b81b4
Use on.exit() for unlinking tempdir
