### 2022-01-10 db27a6a
Use on.exit for segfault test
