### 2022-01-11 82fe3ba
Extend tests for add_resource()
