+++ tests/testthat/test-add_resource.R
@@ -1,19 +1,15 @@
 test_that("add_resource() returns a valid Data Package", {
   pkg <- example_package
-  df <- data.frame(
-    "col_1" = c(1, 2),
-    "col_2" = factor(c("a", "b"), levels = c("a", "b", "c"))
-  )
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
+  df_csv <- "data/df.csv"
   schema <- create_schema(df)
   expect_true(check_package(add_resource(pkg, "new", df)))
   expect_true(check_package(add_resource(pkg, "new", df, schema)))
+  expect_true(check_package(add_resource(pkg, "new", df_csv)))
 })
 
 test_that("add_resource() returns error on incorrect Data Package", {
-  df <- data.frame(
-    "col_1" = c(1, 2),
-    "col_2" = factor(c("a", "b"), levels = c("a", "b", "c"))
-  )
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   expect_error(
     add_resource(list(), "new", df),
     "`package` must be a list object of class `datapackage`",
@@ -24,10 +20,7 @@ test_that("add_resource() returns error on incorrect Data Package", {
 test_that("add_resource() returns error when resource name contains invalid
            characters", {
   pkg <- example_package
-  df <- data.frame(
-    "col_1" = c(1, 2),
-    "col_2" = factor(c("a", "b"), levels = c("a", "b", "c"))
-  )
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   expect_error(
     add_resource(pkg, "New", df),
     paste(
@@ -52,10 +45,7 @@ test_that("add_resource() returns error when resource name contains invalid
 test_that("add_resource() returns error when resource of that name already
            exists", {
   pkg <- example_package
-  df <- data.frame(
-    "col_1" = c(1, 2),
-    "col_2" = factor(c("a", "b"), levels = c("a", "b", "c"))
-  )
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   expect_error(
     add_resource(pkg, "deployments", df),
     "`package` already contains a resource named `deployments`.",
@@ -63,12 +53,19 @@ test_that("add_resource() returns error when resource of that name already
   )
 })
 
+test_that("add_resource() returns error when data is not data frame or
+           character", {
+  pkg <- example_package
+  expect_error(
+    add_resource(pkg, "new", list()),
+    "`data` must be a data frame or path(s) to CSV file(s).",
+    fixed = TRUE
+  )
+})
+
 test_that("add_resource() returns error on invalid or empty data frame", {
   pkg <- example_package
-  df <- data.frame(
-    "col_1" = c(1, 2),
-    "col_2" = factor(c("a", "b"), levels = c("a", "b", "c"))
-  )
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   schema <- create_schema(df)
   expect_error(
     add_resource(pkg, "new", data.frame("col_1" = character(0))),
@@ -84,16 +81,64 @@ test_that("add_resource() returns error on invalid or empty data frame", {
   # For more tests see test-check_schema.R
 })
 
-test_that("add_resource() returns error on incorrect Table Schema", {
+test_that("add_resource() returns error if CSV file cannot be found", {
   pkg <- example_package
-  df <- data.frame(
-    "col_1" = c(1, 2),
-    "col_2" = factor(c("a", "b"), levels = c("a", "b", "c"))
[…]
