### 2022-01-11 0e53384
Only support delim parameter, not "..."
Passing non-default parameters to read_delim() works to assess the file in add_resource(), but the resulting file won't be readable by read_resource(), since none of these parameters are translated to the CSV dialect.

Rather than support all possible parameters, I'm only supporting the main deviating one: the delimiter.
