### 2022-01-11 9a5b00c
Fix #77 and download files if at least one path is local
Done with property "read_from="url" if all paths are URLs. If even one element in the path array is a local file (e.g. mixed among URLs), all files will be copied or downloaded. Streamlining as such makes sense and also makes the implementation easier, because the "read_from" property is assigned for the whole resource, not each path in the path array.
