+++ R/get_resource.R
@@ -43,10 +43,14 @@ get_resource <- function(package, resource_name) {
 
   # Assign read_from property (based on path, then df, then data)
   if (length(resource$path) != 0) {
-    resource$read_from <- "path"
+    if (all(startsWith(resource$path, "http"))) {
+      resource$read_from <- "url"
+    } else {
+      resource$read_from <- "path"
+    }
+    # Expand paths to full paths, check if file exists and check path safety,
+    # unless those paths were willingly added by user in add_resource()
     if (replace_null(attributes(resource)$path, "") != "added") {
-      # Check safety of paths and update to full paths, unless those paths were
-      # willingly added by the user in add_resource().
       resource$path <- purrr::map_chr(
         resource$path, ~ check_path(.x, package$directory, unsafe = FALSE)
       )
