+++ DESCRIPTION
@@ -25,7 +25,8 @@ Imports:
     jsonlite,
     purrr,
     readr (>= 2.1.0),
-    stringr
+    stringr,
+    utils
 Suggests:
     knitr,
     hms,
