### 2022-01-11 b1e55a2
Return output rather than input package invisibly in write_package()
Allows to assess properties in tests without needing to use read_package()
