### 2022-01-11 51ba2be
Solve #65
Don't remove the tempdir, use a subdir
