+++ tests/testthat/test-add_resource.R
@@ -264,14 +264,12 @@ test_that("add_resource() can add resource from CSV file with other delimiter,
   expect_identical(read_resource(pkg, "df_delim"), read_resource(pkg, "df"))
 })
 
-if (FALSE) {
-  test_that("add_resource() creates resource that can be passed to
-             write_package()", {
-    pkg <- example_package
-    df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
-    pkg <- add_resource(pkg, "new", df)
-    temp_dir <- tempdir()
-    on.exit(unlink(temp_dir, recursive = TRUE))
-    expect_invisible(write_package(pkg, temp_dir)) # Can write successfully
-  })
-}
+test_that("add_resource() creates resource that can be passed to
+           write_package()", {
+  pkg <- example_package
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
+  pkg <- add_resource(pkg, "new", df)
+  dir <- file.path(tempdir(), "package")
+  on.exit(unlink(dir, recursive = TRUE))
+  expect_invisible(write_package(pkg, dir)) # Can write successfully
+})
