### 2022-01-11 55e5230
Silence download file messages
Annoying in tests and cannot be silenced with suppressMessages(), see https://stackoverflow.com/questions/66138345/how-to-suppress-download-file-trying-url-message-in-r
