### 2022-01-12 5821ae6
Use "p" as shorthand for package, not pkg
pkg should be reserved for R packages
Update some lines so all fits within 80 char
