+++ tests/testthat/test-remove_resource.R
@@ -1,6 +1,6 @@
 test_that("remove_resource() returns a valid Data Package", {
-  pkg <- example_package
-  expect_true(check_package(remove_resource(pkg, "deployments")))
+  p <- example_package
+  expect_true(check_package(remove_resource(p, "deployments")))
 })
 
 test_that("remove_resource() returns error on incorrect Data Package", {
@@ -15,9 +15,9 @@ test_that("remove_resource() returns error on incorrect Data Package", {
 })
 
 test_that("remove_resource() returns error when resource not found", {
-  pkg <- example_package
+  p <- example_package
   expect_error(
-    remove_resource(pkg, "no_such_resource"),
+    remove_resource(p, "no_such_resource"),
     paste(
       "Can't find resource `no_such_resource` in `deployments`,",
       "`observations`, `media`."
@@ -27,15 +27,15 @@ test_that("remove_resource() returns error when resource not found", {
 })
 
 test_that("remove_resource() removes resource, resource_name", {
-  pkg <- example_package
+  p <- example_package
 
   # Remove "deployments", keep "observations" and "media
-  pkg_removed <- remove_resource(pkg, "deployments")
+  p_removed <- remove_resource(p, "deployments")
 
   # Resource removed
-  expect_length(pkg_removed$resources, 2) # Remains a list, now of length 2
-  expect_identical(pkg_removed$resources[[1]][["name"]], "observations")
+  expect_length(p_removed$resources, 2) # Remains a list, now of length 2
+  expect_identical(p_removed$resources[[1]][["name"]], "observations")
 
   # Resource name removed
-  expect_identical(pkg_removed$resource_names, c("observations", "media"))
+  expect_identical(p_removed$resource_names, c("observations", "media"))
 })
