+++ tests/testthat/test-read_resource.R
@@ -1,14 +1,11 @@
 test_that("read_resource() returns a tibble", {
-  pkg <- example_package
-  df <- data.frame(
-    "col_1" = c(1, 2),
-    "col_2" = factor(c("a", "b"), levels = c("a", "b", "c"))
-  )
-  pkg <- add_resource(pkg, "new", df)
+  p <- example_package
+  df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
+  p <- add_resource(p, "new", df)
 
-  expect_s3_class(read_resource(pkg, "deployments"), "tbl") # via path
-  expect_s3_class(read_resource(pkg, "media"), "tbl")       # via data
-  expect_s3_class(read_resource(pkg, "new"), "tbl")         # via df
+  expect_s3_class(read_resource(p, "deployments"), "tbl") # via path
+  expect_s3_class(read_resource(p, "media"), "tbl")       # via data
+  expect_s3_class(read_resource(p, "new"), "tbl")         # via df
 })
 
 test_that("read_resource() returns error on incorrect Data Package", {
@@ -23,11 +20,11 @@ test_that("read_resource() returns error on incorrect Data Package", {
 })
 
 test_that("read_resource() returns error on incorrect resource", {
-  pkg <- example_package
+  p <- example_package
 
   # No such resource
   expect_error(
-    read_resource(pkg, "no_such_resource"),
+    read_resource(p, "no_such_resource"),
     paste(
       "Can't find resource `no_such_resource` in `deployments`,",
       "`observations`, `media`."
@@ -36,53 +33,56 @@ test_that("read_resource() returns error on incorrect resource", {
   )
 
   # Create invalid package and add properties one by one to pass errors
-  pkg_invalid <- create_package()
-  pkg_invalid$resources <- list(list(name = "deployments"))
-  pkg_invalid$resource_names <- c("deployments")
+  p_invalid <- create_package()
+  p_invalid$resources <- list(list(name = "deployments"))
+  p_invalid$resource_names <- c("deployments")
 
   # No path or data
   expect_error(
-    read_resource(pkg_invalid, "deployments"),
+    read_resource(p_invalid, "deployments"),
     "Resource `deployments` must have property `path` or `data`.",
     fixed = TRUE
   )
 
   # No file at path url
-  pkg_invalid$resources[[1]]$path <- "http://example.com/no_such_file.csv"
+  p_invalid$resources[[1]]$path <- "http://example.com/no_such_file.csv"
   expect_error(
-    read_resource(pkg_invalid, "deployments"),
+    read_resource(p_invalid, "deployments"),
     "Can't find file at `http://example.com/no_such_file.csv`.",
     fixed = TRUE
   )
 
   # No file at path
-  pkg_invalid$resources[[1]]$path <- "no_such_file.csv"
+  p_invalid$resources[[1]]$path <- "no_such_file.csv"
   expect_error(
-    read_resource(pkg_invalid, "deployments"),
+    read_resource(p_invalid, "deployments"),
     "Can't find file at `./no_such_file.csv`.",
     fixed = TRUE
   )
 
   # No file at paths
-  pkg_invalid$resources[[1]]$path <- c("deployments.csv", "no_file.csv")
+  p_invalid$resources[[1]]$path <- c("deployments.csv", "no_such_file.csv")
   expect_error(
-    read_resource(pkg_invalid, "deployments"),
+    read_resource(p_invalid, "deployments"),
     "Can't find file at `./deployments.csv`.",
     fixed = TRUE
   )
 
   # Path is absolute path
-  pkg_invalid$resources[[1]]$path <- "/inst/extdata/deployments.csv"
+  p_invalid$resources[[1]]$path <- "/inst/extdata/deployments.csv"
   expect_error(
[…]
