+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.1
 pkgdown_sha: ~
 articles:
   frictionless: frictionless.html
-last_built: 2022-01-06T17:32Z
+last_built: 2022-01-12T10:09Z
 urls:
   reference: https://frictionlessdata.github.io/frictionless-r/reference
   article: https://frictionlessdata.github.io/frictionless-r/articles
