### 2022-01-12 9dfce1a
Update example_package so that observations resource has URL path
