### 2022-01-12 0ae1182
Rewrite tests for write_package, to capture all resource types
