+++ tests/testthat/test-write_package.R
@@ -1,16 +1,17 @@
 test_that("write_package() returns output Data Package (invisibly)", {
-  pkg <- example_package
+  p <- example_package
+  # Note write_package() is expected to create directory without warning
   dir <- file.path(tempdir(), "package")
   on.exit(unlink(dir, recursive = TRUE))
-  pkg_returned <- write_package(pkg, dir)
-  pkg_as_written <- suppressMessages(read_package(
+  p_written <- write_package(p, dir)
+  p_from_file <- suppressMessages(read_package(
     file.path(dir, "datapackage.json")
   ))
-  # pkg_as_written$directory will differ: overwrite to make the same
-  pkg_as_written$directory <- pkg_returned$directory
+  # p_from_file$directory will differ: overwrite to make the same
+  p_from_file$directory <- p_written$directory
 
-  expect_invisible(write_package(pkg, dir))
-  expect_identical(pkg_returned, pkg_as_written)
+  expect_invisible(write_package(p, dir))
+  expect_identical(p_written, p_from_file)
 })
 
 test_that("write_package() returns error on incorrect Data Package", {
@@ -25,11 +26,11 @@ test_that("write_package() returns error on incorrect Data Package", {
 })
 
 test_that("write_package() returns error if Data Package has no resource(s)", {
-  pkg_empty <- create_package()
-  temp_dir <- tempdir()
-  on.exit(unlink(temp_dir, recursive = TRUE))
+  p_empty <- create_package()
+  dir <- file.path(tempdir(), "package")
+  on.exit(unlink(dir, recursive = TRUE))
   expect_error(
-    write_package(pkg_empty, temp_dir),
+    write_package(p_empty, dir),
     "`package` must have resources. Use `add_resource()` to add resources.",
     fixed = TRUE
   )
@@ -38,170 +39,205 @@ test_that("write_package() returns error if Data Package has no resource(s)", {
   # Resources without path or data are tested in test-read_resource.R
 })
 
-test_that("write_package() writes to the specified directory", {
-  pkg <- example_package
-  temp_subdir <- file.path(tempdir(), "x/y")
-  on.exit(unlink(temp_subdir, recursive = TRUE))
-
-  # Function should create subdir(s) without error
-  expect_invisible(write_package(pkg, temp_subdir))
-
-  # Valid package can be found at temp_subdir
-  expect_true(check_package(
-    suppressMessages(read_package(file.path(temp_subdir, "datapackage.json")))
-  ))
-})
-
 test_that("write_package() writes unaltered datapackage.json as is", {
-  pkg <- suppressMessages(read_package(
+  p <- suppressMessages(read_package(
     system.file("extdata", "datapackage.json", package = "frictionless")
   ))
-  json_in <- readr::read_file(
+  json_original <- readr::read_file(
     system.file("extdata", "datapackage.json", package = "frictionless")
   )
-  temp_dir <- tempdir()
-  on.exit(unlink(temp_dir, recursive = TRUE))
-  write_package(pkg, temp_dir)
-  json_out <- readr::read_file(file.path(temp_dir, "datapackage.json"))
+  dir <- file.path(tempdir(), "package")
+  on.exit(unlink(dir, recursive = TRUE))
+  write_package(p, dir)
+  json_as_written <- readr::read_file(file.path(dir, "datapackage.json"))
 
   # Output json = input json. This also tests if new properties (resource_names,
   # directories) are removed and json is printed "pretty"
-  expect_identical(json_out, json_in)
+  expect_identical(json_as_written, json_original)
 })
 
[…]
