### 2022-01-12 #495 @peterdesmet
Hi @melvidoni, 2 questions:
1. I'm about to merge a PR with updated functionality into the package. Would it be ok if the reviewers review the resulting `0.10.0` version?
