### 2022-01-12 e7f01e8
Explain why some files are not written in example
