### 2022-01-12 39368d9
Add NEWS.md
