### 2022-01-12 5d01775
Minor updates in documentation
