+++ R/add_resource.R
@@ -44,14 +44,14 @@
 #'   y = c(860, 900)
 #' )
 #'
-#' # Add a resource "positions" to the Data Package, from the data frame
+#' # Add resource "positions" to the Data Package, from the data frame
 #' package <- add_resource(package, "positions", data = df)
 #'
-#' # Add a resource "positions2" to the Data Package, with user-defined schema
+#' # Add resource "positions2" to the Data Package, with user-defined schema
 #' my_schema <- create_schema(df)
 #' package <- add_resource(package, "positions2", data = df, schema = my_schema)
 #'
-#' # Add a resource "observations2" to the Data Package, from CSV file paths
+#' # Add resource "observations2" to the Data Package, from CSV file paths
 #' path_1 <- system.file("extdata", "observations_1.csv", package = "frictionless")
 #' path_2 <- system.file("extdata", "observations_2.csv", package = "frictionless")
 #' package <- add_resource(package, "deployments2", data = c(path_1, path_2))
