+++ R/read_resource.R
@@ -187,7 +187,7 @@
 #' # The above tibble is merged from 2 files listed in the resource path
 #' package$resources[[2]]$path
 #'
-#' # With col_names and col_types derived from the resource schema
+#' # The column names and types are derived from the resource schema
 #' purrr::map_chr(package$resources[[2]]$schema$fields, "name")
 #' purrr::map_chr(package$resources[[2]]$schema$fields, "type")
 read_resource <- function(package, resource_name) {
