+++ R/create_schema.R
@@ -62,7 +62,7 @@
 #'   life_stage = factor(c("adult", "adult"), levels = c("adult", "juvenile"))
 #' )
 #'
-#' # Create a Table Schema
+#' # Create a Table Schema from the data frame
 #' schema <- create_schema(df)
 #' str(schema)
 create_schema <- function(data) {
