+++ man/read_resource.Rd
@@ -209,7 +209,7 @@ read_resource(package, "observations")
 # The above tibble is merged from 2 files listed in the resource path
 package$resources[[2]]$path
 
-# With col_names and col_types derived from the resource schema
+# The column names and types are derived from the resource schema
 purrr::map_chr(package$resources[[2]]$schema$fields, "name")
 purrr::map_chr(package$resources[[2]]$schema$fields, "type")
 }
