### 2022-01-12 4c2d1b4
Show message when downloading file (suppress in tests)
