### 2022-01-13 46afd1a
Set columns containing NA only to "string", fix #79
Alternative is something like:

dplyr::mutate(df, dplyr::across(tidyselect::where(all(is.na)), as.character))

but that requires tidyselect and it will raise an error for col names that are NA (allowed, there is a test for that)
