+++ R/create_schema.R
@@ -76,6 +76,16 @@ create_schema <- function(data) {
     )
   )
 
+  # Columns with all NA are considered logical by R (and read_delim)
+  # Set those to character, since string is a better default for Table Schema
+  data_as_list <- lapply(data, function(x) {
+    if (all(is.na(x))) {
+      as.character(x)
+    } else {
+      x
+    }
+  })
+
   # Create fields (a list of lists)
   fields <- purrr::imap(data_as_list, function(x, name) {
     # Name
