### 2022-01-13 df6ce93
Only convert to NA if originally interpreted as logical
