### 2022-01-13 c624515
Describe updated create_schema() behaviour in #79
