### 2022-01-14 cfdc0f4
Use "safe" rather "unsafe" in check_path() + call explicitely
