### 2022-01-14 28343f8
Require dialect (and schema, already implemented) to be safe path
