### 2022-01-14 1ac6297
Update read_json to read_descriptor and support yaml for schema/dialect
