+++ tests/testthat/data/deployments_schema.yaml
@@ -1,0 +1,29 @@
+fields:
+- name: deployment_id
+  type: string
+  constraints:
+    required: true
+    unique: true
+- name: longitude
+  type: number
+  constraints:
+    required: true
+    minimum: -180
+    maximum: 180
+- name: latitude
+  constraints:
+    required: true
+- name: start
+  type: date
+  format: "%x"
+  constraints:
+    required: true
+- name: comments
+  type: string
+  constraints:
+    required: false
+missingValues:
+- ''
+- NA
+- NaN
+primaryKey: deployment_id
