### 2022-01-14 3dd0adc
Use read_descriptor for read_package and test input
