### 2022-01-14 2842691
Update NEWS.md
