### 2022-01-14 01b487e
Support descriptor, schema, dialect in YAML, fix #66 (#82)
* Use "safe" rather "unsafe" in check_path() + call explicitely

* Require dialect (and schema, already implemented) to be safe path

* Update read_json to read_descriptor and support yaml for schema/dialect

* Use read_descriptor for read_package and test input

* Update NEWS.md
