+++ tests/testthat/test-read_resource.R
@@ -130,30 +130,6 @@ test_that("read_resource() returns error on incorrect resource", {
     # no fixed = TRUE, since full returned path depends on system
   )
 
-  # Schema is absolute path
-  p_invalid$resources[[1]]$schema <- file.path(
-    "/tests/testthat/data/deployments_schema.json"
-  )
-  expect_error(
-    read_resource(p_invalid, "deployments"),
-    paste(
-      "`/tests/testthat/data/deployments_schema.json` is an absolute path",
-      "(`/`) which is unsafe."
-    ),
-    fixed = TRUE
-  )
-
-  # Schema is relative parent path
-  p_invalid$resources[[1]]$schema <- "../testthat/data/deployments_schema.json"
-  expect_error(
-    read_resource(p_invalid, "deployments"),
-    paste(
-      "`../testthat/data/deployments_schema.json` is a relative parent path",
-      "(`../`) which is unsafe."
-    ),
-    fixed = TRUE
-  )
-
   # No fields
   p_invalid$resources[[1]]$schema <- list()
   expect_error(
@@ -221,48 +197,111 @@ test_that("read_resource() can read remote files", {
   expect_identical(read_resource(p_remote_resource, "deployments"), resource)
 })
 
-test_that("read_resource() can read local and remote Table Schemas", {
+test_that("read_resource() can read safe local and remote Table Schema,
+           including YAML", {
   p <- example_package
   resource <- read_resource(p, "deployments")
-
-  p_local_schema <- p
-  p_local_schema$directory <- "." # Use "./tests/testthat" outside test
-  p_local_schema$resources[[1]]$schema <- "data/deployments_schema.json"
-  # Using a remote path, otherwise schema and path need to share same directory
-  p_local_schema$resources[[1]]$path <- file.path(
+  p$directory <- "."
+  # Use a remote path, otherwise schema and path need to share same directory
+  p$resources[[1]]$path <- file.path(
     "https://github.com/frictionlessdata/frictionless-r",
     "raw/main/inst/extdata/deployments.csv"
   )
+
+  # Schema is absolute path
+  p_unsafe <- p
+  p_unsafe$resources[[1]]$schema <- file.path(
+    "/tests/testthat/data/deployments_schema.json"
+  )
+  expect_error(
+    read_resource(p_unsafe, "deployments"),
+    paste(
+      "`/tests/testthat/data/deployments_schema.json` is an absolute path",
+      "(`/`) which is unsafe."
+    ),
+    fixed = TRUE
+  )
+
+  # Schema is relative parent path
+  p_unsafe$resources[[1]]$schema <- "../testthat/data/deployments_schema.json"
+  expect_error(
+    read_resource(p_unsafe, "deployments"),
+    paste(
+      "`../testthat/data/deployments_schema.json` is a relative parent path",
+      "(`../`) which is unsafe."
+    ),
+    fixed = TRUE
+  )
+
+  # Schema is local path
+  p_local_schema <- p
+  p_local_schema$resources[[1]]$schema <- "data/deployments_schema.json"
   expect_identical(read_resource(p_local_schema, "deployments"), resource)
 
+  # Schema is remote path
   p_remote_schema <- p
   p_remote_schema$resources[[1]]$schema <- file.path(
     "https://github.com/frictionlessdata/frictionless-r",
     "raw/main/tests/testthat/data/deployments_schema.json"
   )
   expect_identical(read_resource(p_remote_schema, "deployments"), resource)
+
+  # Schema is YAML
+  p_yaml_schema <- p
+  p_yaml_schema$resource[[1]]$schema <- "data/deployment_schema.yaml"
[…]
