### 2022-01-14 5dec7a4
Create functions for utils functions
Except for check_path() and read_descriptor() which are sufficiently tested in main functions.
