+++ DESCRIPTION
@@ -29,11 +29,11 @@ Imports:
     utils,
     yaml
 Suggests:
-    knitr,
     hms,
+    knitr,
     lubridate,
-    testthat (>= 3.0.0),
-    rmarkdown
+    rmarkdown,
+    testthat (>= 3.0.0)
 Config/testthat/edition: 3
 Encoding: UTF-8
 LazyData: true
