### 2022-01-17 a33e140
Update `unique_sorted()` (#84)
* Create functions for utils functions, except for check_path() and read_descriptor() which are sufficiently tested in main functions.

* Rewrite unique_sorted() without dplyr

* Order Suggests alphabetically

* Update codemeta.json
