+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 2.7.3
-pkgdown: 2.0.1
+pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   frictionless: frictionless.html
-last_built: 2022-01-14T14:47Z
+last_built: 2022-01-17T11:00Z
 urls:
   reference: https://frictionlessdata.github.io/frictionless-r/reference
   article: https://frictionlessdata.github.io/frictionless-r/articles
