### 2022-01-17 #495 @peterdesmet
@melvidoni @zambujo Thanks!

Version `0.10.0` of the package has just been [released](https://github.com/frictionlessdata/frictionless-r/tree/v0.10.0), which would be the preferred version for review.
