### 2022-01-17 755e362
Set `format`, `mediatype` and `encoding` for added CSV resources
Fix #78
