### 2022-01-17 #495 @peterdesmet
@melvidoni version 0.10.0 has been merged to the default branch (`main`), but that branch is also used for further development.

To install `0.10.0` specifically (recommended):

```R
devtools::install_github("frictionlessdata/frictionless-r@v0.10.0")
```

To install the latest development version (`0.10.0.9000`):

```R
devtools::install_github("frictionlessdata/frictionless-r")
```
