### 2022-01-18 f5ac397
Move decimalChar/groupChar test closer to encoding
