### 2022-01-18 4575771
Shorten decimal/grouping mark messages and test for entire message
