### 2022-01-18 fc240b7
Make sure encoding is known encoding or set to UTF-8
Fix #86
