### 2022-01-18 9cf39df
Set `format`, `mediatype` and `encoding` for added CSV resources (#85)
* Set `format`, `mediatype` and `encoding` for added CSV resources, fix #78

* Update news

* Update heading in news

* Use colon
