+++ tests/testthat/test-add_resource.R
@@ -180,7 +180,7 @@ test_that("add_resource() adds resource, resource_name", {
 })
 
 test_that("add_resource() uses provided schema or creates one", {
-  p <- example_package
+  p <- create_package()
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
   df_csv <- "data/df.csv"
   schema <- create_schema(df)
@@ -192,16 +192,16 @@ test_that("add_resource() uses provided schema or creates one", {
   # df
   p <- add_resource(p, "new_df", df)
   p <- add_resource(p, "new_df_with_schema", df, schema_custom)
-  expect_identical(p$resources[[4]]$schema, schema)
-  expect_identical(p$resources[[5]]$schema, schema_custom)
+  expect_identical(p$resources[[1]]$schema, schema)
+  expect_identical(p$resources[[2]]$schema, schema_custom)
   expect_identical(get_schema(p, "new_df"), schema)
   expect_identical(get_schema(p, "new_df_with_schema"), schema_custom)
 
   # csv
   p <- add_resource(p, "new_csv", df)
   p <- add_resource(p, "new_csv_with_schema", df, schema_custom)
-  expect_identical(p$resources[[6]]$schema, schema)
-  expect_identical(p$resources[[7]]$schema, schema_custom)
+  expect_identical(p$resources[[3]]$schema, schema)
+  expect_identical(p$resources[[4]]$schema, schema_custom)
   expect_identical(get_schema(p, "new_csv"), schema)
   expect_identical(get_schema(p, "new_csv_with_schema"), schema_custom)
 })
@@ -214,8 +214,8 @@ test_that("add_resource() can add resource from data frame, readable by
   expect_identical(read_resource(p, "new"), dplyr::as_tibble(df))
 })
 
-test_that("add_resource() can add resource from local, remote, relative,
-           absolute or compressed CSV file, readable by read_resource()", {
+test_that("add_resource() can add resource from local, relative, absolute,
+           remote or compressed CSV file, readable by read_resource()", {
   p <- example_package
   schema <- get_schema(p, "deployments")
 
@@ -225,19 +225,10 @@ test_that("add_resource() can add resource from local, remote, relative,
   expect_identical(p$resources[[4]]$path, local_path)
   expect_s3_class(read_resource(p, "local"), "tbl")
 
-  # Remote
-  remote_path <- file.path(
-    "https://github.com/frictionlessdata/frictionless-r",
-    "raw/main/inst/extdata/deployments.csv"
-  )
-  p <- add_resource(p, "remote", remote_path, schema)
-  expect_identical(p$resources[[5]]$path, remote_path)
-  expect_s3_class(read_resource(p, "remote"), "tbl")
-
   # Relative (doesn't throw unsafe error)
   relative_path <- "../testthat/data/df.csv"
   p <- add_resource(p, "relative", relative_path)
-  expect_identical(p$resources[[6]]$path, relative_path)
+  expect_identical(p$resources[[5]]$path, relative_path)
   expect_s3_class(read_resource(p, "relative"), "tbl")
 
   # Absolute (doesn't throw unsafe error)
@@ -245,8 +236,17 @@ test_that("add_resource() can add resource from local, remote, relative,
     "extdata", "deployments.csv", package = "frictionless" # Will start with /
   )
   p <- add_resource(p, "absolute", absolute_path, schema)
-  expect_identical(p$resources[[7]]$path, absolute_path)
[…]
