### 2022-01-18 1ed3c05
Warn on unknown encodings and read with UTF-8 (#87)
* Move decimalChar/groupChar test closer to encoding

* Shorten decimal/grouping mark messages and test for entire message

* Make sure encoding is known encoding or set to UTF-8 Fix #86
