+++ tests/testthat/test-read_resource.R
@@ -350,12 +350,74 @@ test_that("read_resource() understands encoding", {
   p <- example_package
   resource <- read_resource(p, "deployments")
 
-  # Create package with non-default missing values
+  # Create package with non-default encoding
   p_encoding <- p
   p_encoding$directory <- "."
   p_encoding$resources[[1]]$path <- "data/deployments_encoding.csv"
   p_encoding$resources[[1]]$encoding <- "windows-1252"
   expect_identical(read_resource(p_encoding, "deployments"), resource)
+
+  # Create package with unknown encoding
+  p_unknown <- p
+  p_unknown$resources[[1]]$encoding <- "utf-8-sig"
+  warnings <- capture_warnings(read_resource(p_unknown, "deployments"))
+  expect_identical(
+    warnings[1],
+    "Unknown encoding `utf-8-sig`. Reading file(s) with UTF-8 encoding."
+  )
+  expect_identical(
+    suppressWarnings(read_resource(p_unknown, "deployments")),
+    resource
+  )
+})
+
+test_that("read_resource() handles decimalChar/groupChar properties", {
+  expected_value <- 3000000.3
+  p <- suppressMessages(read_package("data/mark.json"))
+
+  # Default decimalChar/groupChar
+  resource <- read_resource(p, "mark_default")
+  expect_identical(resource$num, expected_value) # 3000000.30
+  expect_identical(resource$num_undefined, expected_value) # 3000000.30
+
+  # Non-default decimalChar, default groupChar (which should not conflict)
+  warnings <- capture_warnings(read_resource(p, "mark_decimal"))
+  expect_identical(
+    warnings[1],
+    paste(
+      "Some fields define a non-default `decimalChar`.",
+      "Parsing all number fields with `,` as decimal mark."
+    )
+  )
+
+  resource <- suppressWarnings(read_resource(p, "mark_decimal"))
+  expect_identical(resource$num, expected_value) # 3000000.30
+  expect_identical(resource$num_undefined, expected_value) # 3000000.30
+
+  # Non-default decimalChar/groupChar
+  warnings <- capture_warnings(read_resource(p, "mark_decimal_group"))
+  expect_true(length(warnings) == 3) # 2 warnings + 1 parsing failure last field
+  expect_identical(
+    warnings[1],
+    paste(
+      "Some fields define a non-default `decimalChar`.",
+      "Parsing all number fields with `,` as decimal mark."
+    )
+  )
+  expect_identical(
+    warnings[2],
+    paste(
+      "Some fields define a non-default `groupChar`.",
+      "Parsing all number fields with `.` as grouping mark."
+    )
+  )
+
+  resource <- suppressWarnings(read_resource(p, "mark_decimal_group"))
+  expect_identical(resource$num, expected_value) # 3.000.000,30
+  # Field without decimalChar is still parsed with non-default decimalChar
+  expect_identical(resource$num_undefined, expected_value) # 3000000,30
+  # Field without groupChar is not parsed with non-default groupChar
+  expect_identical(resource$num_undefined_group, NA_real_) # 3.000.000,30
 })
 
 test_that("read_resource() handles LF and CRLF line terminator characters", {
@@ -584,35 +646,4 @@ test_that("read_resource() handles other types", {
 
   # Guess undefined types, unknown types are blocked by check_schema()
   expect_type(resource$no_type, "logical")
-})
-
-test_that("read_resource() handles decimalChar/groupChar properties", {
-  expected_value <- 3000000.3
-  p <- suppressMessages(read_package("data/mark.json"))
-
-  # Default decimalChar/groupChar
-  resource <- read_resource(p, "mark_default")
-  expect_identical(resource$num, expected_value) # 3000000.30
[…]
