+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - `add_resource()` now sets `format`, `mediatype` and `encoding` for added CSV 
   file(s) (#78).
+- `read_resource()` will now warn rather than error on unknown encoding (#86).
 
 # frictionless 0.10.0
 
