### 2022-01-19 2ca5808
Don't require datapackage attribute + update error message
