+++ R/write_resource.R
@@ -4,8 +4,8 @@
 #'
 #' @inheritParams read_resource
 #' @param directory Path to local directory to write files to.
-#' @return Updated list object describing a Data Resource, ready to be included
-#'   in a `datapackage.json`.
+#' @return Updated list describing a Data Resource, ready to be included in a
+#'   `datapackage.json`.
 #' @family write functions
 #' @noRd
 write_resource <- function(package, resource_name, directory = ".") {
