+++ tests/testthat/test-write_package.R
@@ -18,8 +18,8 @@ test_that("write_package() returns error on incorrect Data Package", {
   expect_error(
     write_package(list()),
     paste(
-      "`package` must be a list object created with `read_package()` or",
-      "`create_package()`."
+      "`package` must be a list describing a Data Package,",
+      "created with `read_package()` or `create_package()`."
     ),
     fixed = TRUE
   )
