+++ tests/testthat/test-add_resource.R
@@ -13,8 +13,8 @@ test_that("add_resource() returns error on incorrect Data Package", {
   expect_error(
     add_resource(list(), "new", df),
     paste(
-      "`package` must be a list object created with `read_package()` or",
-      "`create_package()`."
+      "`package` must be a list describing a Data Package,",
+      "created with `read_package()` or `create_package()`."
     ),
     fixed = TRUE
   )
