### 2022-01-19 68f307e
Update NEWS.md
