### 2022-01-19 50543ec
Add purrr::prepend() example in vignette
