### 2022-01-19 9ea4957
Drop `datapackage` attribute, making it easier to use `append()` to add attributes. (#89)
* Don't add datapackage attribute

* Don't require datapackage attribute + update error message

* Shorten "list object" to simply "list"

* Update NEWS.md

* Add purrr::prepend() example in vignette

* Use base `append()` rather `purrr::prepend()`

* Remove custom attribute from example_package
