+++ tests/testthat/test-check_package.R
@@ -9,44 +9,35 @@ test_that("check_package() returns error on incorrect Data Package", {
     resource_names = vector(mode = "character"),
     directory = "."
   )
-  class(p) <- c("datapackage", class(p))
   expect_true(check_package(p))
 
-  # Must be a list
-  expect_error(
-    check_package("not_a_list"),
-    paste(
-      "`package` must be a list object of class `datapackage` created with",
-      "`read_package()` or `create_package()`."
-    ),
-    fixed = TRUE
+  error_message <- paste(
+    "`package` must be a list describing a Data Package,",
+    "created with `read_package()` or `create_package()`."
   )
-
-  # Must have class datapackage
-  p_invalid <- p
-  class(p_invalid) <- "list"
-  expect_error(check_package(p_invalid), "`package` must be")
+  # Must be a list
+  expect_error(check_package("not_a_list"), error_message, fixed = TRUE)
 
   # Must have resources as list
   p_invalid <- p
   p_invalid$resources <- NULL
-  expect_error(check_package(p_invalid), "`package` must be")
+  expect_error(check_package(p_invalid), error_message, fixed = TRUE)
   p_invalid$resources <- vector(mode = "character")
-  expect_error(check_package(p_invalid), "`package` must be")
+  expect_error(check_package(p_invalid), error_message, fixed = TRUE)
 
   # Must have resource names as character vector
   p_invalid <- p
   p_invalid$resource_names <- NULL
-  expect_error(check_package(p_invalid), "`package` must be")
+  expect_error(check_package(p_invalid), error_message, fixed = TRUE)
   p_invalid$resource_names <- list()
-  expect_error(check_package(p_invalid), "`package` must be")
+  expect_error(check_package(p_invalid), error_message, fixed = TRUE)
 
   # Must have directory as character
   p_invalid <- p
   p_invalid$directory <- NULL
-  expect_error(check_package(p_invalid), "`package` must be")
+  expect_error(check_package(p_invalid), error_message, fixed = TRUE)
   p_invalid$directory <- logical()
-  expect_error(check_package(p_invalid), "`package` must be")
+  expect_error(check_package(p_invalid), error_message, fixed = TRUE)
 })
 
 test_that("check_package() returns error if resources have no name", {
