+++ R/get_resource.R
@@ -4,8 +4,8 @@
 #' from a Data Package, i.e. the content of one of the described `resources`.
 #'
 #' @inheritParams read_resource
-#' @return List object describing a Data Resource, with new property `read_from`
-#'   to indicate how data should be read.
+#' @return List describing a Data Resource, with new property `read_from` to
+#'   indicate how data should be read.
 #'   If present, `path` will be updated to contain the full path(s).
 #' @family edit functions
 #' @noRd
