### 2022-01-19 5761eb4
Use later version of Zenodo example with zipped resources
