+++ README.Rmd
@@ -43,7 +43,7 @@ devtools::install_github("frictionlessdata/frictionless-r")
 
 ## Usage
 
-With frictionless you can directly **read** data from a Data Package (local or remote) into your R environment. Here we read bird GPS tracking data from a Data Package published on [Zenodo](https://doi.org/10.5281/zenodo.5070086):
+With frictionless you can directly **read** data from a Data Package (local or remote) into your R environment. Here we read bird GPS tracking data from a Data Package published on [Zenodo](https://doi.org/10.5281/zenodo.5879096):
 
 ```{r read_example}
 library(frictionless)
@@ -51,14 +51,14 @@ library(frictionless)
 # Read the datapackage.json file
 # This gives you access to all Data Resources of the Data Package without 
 # reading them, which is convenient and fast.
-package <- read_package("https://zenodo.org/record/5070086/files/datapackage.json")
+package <- read_package("https://zenodo.org/record/5879096/files/datapackage.json")
 
 # List the available resources
 package$resource_names
 
 # Read data from the resource "gps"
 # This will return a single data frame, even though the data are split over 
-# multiple CSV files.
+# multiple zipped CSV files.
 read_resource(package, "gps")
 ```
 
