### 2022-01-21 ce1cba2
Support adding a schema via path or URL in add_resource()
