### 2022-01-21 fbfb69e
Support adding a schema via path or URL in add_resource() (#91)
* Support adding a schema via path or URL in add_resource()

* Update NEWS.md

* Use default ~ as homepage (more generic)
