+++ man/example_package.Rd
@@ -5,7 +5,7 @@
 \alias{example_package}
 \title{Example Data Package}
 \format{
-An object of class \code{datapackage} (inherits from \code{list}) of length 9.
+An object of class \code{list} of length 9.
 }
 \source{
 \url{https://github.com/frictionlessdata/frictionless-r/tree/main/inst/extdata}
