+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: https://frictionlessdata.github.io/frictionless-r
+url: ~
 
 template:
   bootstrap: 5
