### 2022-01-24 d2c1b4a
Ignore zip/gz extension + set to csv if no extension is found
