+++ R/add_resource.R
@@ -94,14 +94,20 @@ add_resource <- function(package, resource_name, data, schema = NULL,
     # Check existence of files (no further checks) and read last file
     paths <- purrr::map_chr(data, ~ check_path(.x, safe = FALSE))
     last_file <- utils::tail(paths, n = 1)
-    extension <- utils::tail(strsplit(last_file, "\\.")[[1]], n = 1)
-    encoding <- readr::guess_encoding(last_file, n_max = 1000)[[1, 1]]
-    encoding <- replace_null(encoding, "UTF-8")
     df <- readr::read_delim(
-      file = paths[length(paths)],
+      file = last_file,
       delim = delim,
       show_col_types = FALSE,
     )
+
+    # Get extension, inspired by tools::file_ext()
+    file_name <- gsub("(\\.gz$)|(\\.zip$)", "", last_file) # Ignore compression
+    pos <- regexpr("\\.([[:alnum:]]+)$", file_name)
+    extension <- ifelse(pos > -1L, substring(file_name, pos + 1L), "csv")
+
+    # Get encoding
+    encoding <- readr::guess_encoding(last_file, n_max = 1000)[[1, 1]]
+    encoding <- replace_null(encoding, "UTF-8")
   }
 
   # Create schema
