### 2022-01-24 ea46c6f
Set format based on delimiter: tsv or csv
- Use delimiter, not extension for format, see #90
- Don't use replace_null() for return of guess_encoding(): not necessary
- This reverts commit d2c1b4a500a1cd4df188441dcd712828d27809d2
- Add just one test, for .zip
