### 2022-01-25 4a49dca
Add test for not overwriting existing files (#93)
Fix #88
