### 2022-01-25 c025944
Set format based on delimiter in `add_resource()` (#92)
- Use delimiter, not extension for format, see #90
- Don't use replace_null() for return of guess_encoding(): not necessary
- Add just one test, for .zip
