### 2022-01-25 85d5c2f
Remove stringr dependency
Use switch() rather than str_replace_all() to assign datetime format

Fix #46
