### 2022-01-25 73de59c
Use test_path() so paths work outside tests too
Inspired by https://blog.r-hub.io/2020/05/29/distribute-data/#data-outside-of-your-package
