+++ tests/testthat/test-read_resource.R
@@ -164,7 +164,10 @@ test_that("read_resource() can read newly added data (ignoring schema)", {
 
 test_that("read_resource() can read inline data (ignoring schema)", {
   p <- example_package
-  expected_resource <- readr::read_csv("data/media.csv", col_types = "ccccc")
+  expected_resource <- readr::read_csv(
+    test_path("data/media.csv"),
+    col_types = "ccccc"
+  )
   expect_identical(read_resource(p, "media"), expected_resource)
 
   p$resources[[3]]$data <- "not_a_list" # Media resource
@@ -210,9 +213,8 @@ test_that("read_resource() can read safe local and remote Table Schema,
 
   # Schema is absolute path
   p_unsafe <- p
-  p_unsafe$resources[[1]]$schema <- file.path(
+  p_unsafe$resources[[1]]$schema <-
     "/tests/testthat/data/deployments_schema.json"
-  )
   expect_error(
     read_resource(p_unsafe, "deployments"),
     paste(
@@ -235,7 +237,8 @@ test_that("read_resource() can read safe local and remote Table Schema,
 
   # Schema is local path
   p_local_schema <- p
-  p_local_schema$resources[[1]]$schema <- "data/deployments_schema.json"
+  p_local_schema$resources[[1]]$schema <-
+    test_path("data/deployments_schema.json")
   expect_identical(read_resource(p_local_schema, "deployments"), resource)
 
   # Schema is remote path
@@ -248,7 +251,7 @@ test_that("read_resource() can read safe local and remote Table Schema,
 
   # Schema is YAML
   p_yaml_schema <- p
-  p_yaml_schema$resource[[1]]$schema <- "data/deployment_schema.yaml"
+  p_yaml_schema$resource[[1]]$schema <- test_path("data/deployment_schema.yaml")
   expect_identical(read_resource(p_yaml_schema, "deployments"), resource)
 })
 
@@ -287,7 +290,7 @@ test_that("read_resource() can read safe local and remote CSV dialect", {
 
   # Dialect is local path
   p_local_dialect <- p
-  p_local_dialect$resources[[1]]$dialect <- "data/dialect.json"
+  p_local_dialect$resources[[1]]$dialect <- test_path("data/dialect.json")
   expect_identical(read_resource(p_local_dialect, "deployments"), resource)
 
   # Dialect is remote path
@@ -300,7 +303,7 @@ test_that("read_resource() can read safe local and remote CSV dialect", {
 
   # Dialect is YAML
   p_yaml_dialect <- p
-  p_yaml_dialect$resource[[1]]$dialect <- "data/dialect.yaml"
+  p_yaml_dialect$resource[[1]]$dialect <- test_path("data/dialect.yaml")
   expect_identical(read_resource(p_yaml_dialect, "deployments"), resource)
 })
 
@@ -311,7 +314,7 @@ test_that("read_resource() understands CSV dialect", {
   # Create package with non-default dialect properties
   p_dialect <- p
   p_dialect$directory <- "."
-  p_dialect$resources[[1]]$path <- "data/deployments_dialect.csv"
+  p_dialect$resources[[1]]$path <- test_path("data/deployments_dialect.csv")
   p_dialect$resources[[1]]$dialect <- list(
     delimiter = "/",
     # lineTerminator
@@ -340,7 +343,8 @@ test_that("read_resource() understands missing values", {
   # Create package with non-default missing values
   p_missing <- p
   p_missing$directory <- "."
-  p_missing$resources[[1]]$path <- "data/deployments_missingvalues.csv"
+  p_missing$resources[[1]]$path <-
+    test_path("data/deployments_missingvalues.csv")
   p_missing$resources[[1]]$schema$missingValues <-
[…]
