### 2022-01-25 fa974e2
Use test_path() so paths work outside tests too (#95)
Inspired by https://blog.r-hub.io/2020/05/29/distribute-data/#data-outside-of-your-package
