### 2022-01-25 f05fb68
Remove stringr dependency (#94)
Use switch() rather than str_replace_all() to assign datetime format, fix #46
