### 2022-01-27 499e39d
Add author and fixed date for vignette
Date is set to when last big contribution to vignette was made, rather than `Sys.Date()` which would update with every minor commit.
