+++ vignettes/frictionless.Rmd
@@ -1,5 +1,7 @@
 ---
 title: "Introduction to frictionless"
+author: "Peter Desmet"
+date: 2022-01-03
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Introduction to frictionless}
