### 2022-01-28 1280452
Run write_zenodo_json()
Sorts .Rbuildignore alphabetical
