+++ DESCRIPTION
@@ -6,13 +6,13 @@ Authors@R: c(
            comment = c(ORCID = "0000-0002-8442-8025")),
     person("Damiano", "Oldoni", , "damiano.oldoni@inbo.be", role = "aut",
            comment = c(ORCID = "0000-0003-3445-7562")),
-    person("Research Institute for Nature and Forest (INBO)", , , 
-           "info@inbo.be", role = c("cph"))
+    person("Research Institute for Nature and Forest (INBO)", , , "info@inbo.be", role = "cph")
   )
-Description: Read and write Frictionless Data Packages. A Data Package 
-  (<https://specs.frictionlessdata.io/data-package/>) is a simple container 
-  format and standard to describe and package a collection of (tabular) data. 
-  It is typically used to publish FAIR and open datasets.
+Description: Read and write Frictionless Data Packages. A Data Package
+    (<https://specs.frictionlessdata.io/data-package/>) is a simple
+    container format and standard to describe and package a collection of
+    (tabular) data.  It is typically used to publish FAIR and open
+    datasets.
 License: MIT + file LICENSE
 URL: https://github.com/frictionlessdata/frictionless-r,
     https://frictionlessdata.github.io/frictionless-r/
@@ -33,9 +33,10 @@ Suggests:
     lubridate,
     rmarkdown,
     testthat (>= 3.0.0)
+VignetteBuilder: 
+    knitr
 Config/testthat/edition: 3
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.1.2
-VignetteBuilder: knitr
