+++ README.Rmd
@@ -29,7 +29,7 @@ Frictionless is an R package to read and write Frictionless Data Packages. A [Da
 
 To get started, see:
 
-- [Get started](https://frictionlessdata.github.io/frictionless-r/articles/frictionless.html): an introduction to the package's main functionality.
+- [Get started](https://frictionlessdata.github.io/frictionless-r/articles/frictionless.html): an introduction to the package's main functionalities.
 - [Function reference](https://frictionlessdata.github.io/frictionless-r/reference/index.html): overview of all functions.
 
 ## Installation
@@ -43,7 +43,7 @@ devtools::install_github("frictionlessdata/frictionless-r")
 
 ## Usage
 
-With frictionless you can directly **read** data from a Data Package (local or remote) into your R environment. Here we read bird GPS tracking data from a Data Package published on [Zenodo](https://doi.org/10.5281/zenodo.5879096):
+With frictionless you can **read** data from a Data Package (local or remote) into your R environment. Here we read bird GPS tracking data from a Data Package published on [Zenodo](https://doi.org/10.5281/zenodo.5879096):
 
 ```{r read_example}
 library(frictionless)
@@ -83,7 +83,7 @@ For more functionality, see [get started](https://frictionlessdata.github.io/fri
 
 ## frictionless vs datapackage.r
 
-[datapackage.r](https://cran.r-project.org/web/packages/datapackage.r/) is an alternative R package to work with Data Packages. It has an object-oriented design (using a `Package` class) and offers validation. frictionless on the other hand allows users to quickly read and write Data Package data to and from R data frames, getting out of your way for the rest of your analysis. It is designed to be lightweight, follows [tidyverse](https://www.tidyverse.org/) principles and supports piping.
+[datapackage.r](https://cran.r-project.org/web/packages/datapackage.r/) is an alternative R package to work with Data Packages. It has an object-oriented design (using a `Package` class) and offers validation. frictionless on the other hand allows you to quickly read and write Data Packages to and from data frames, getting out of the your for the rest of your analysis. It is designed to be lightweight, follows [tidyverse](https://www.tidyverse.org/) principles and supports piping.
 
 ## Meta
 
