### 2022-01-31 #495 @beatrizmilz
Hi! Thanks for inviting. Here is my review:

Estimated hours spent reviewing: 3h

---

**Congratulations to the authors. Everything worked well. The comments I wrote were mostly ideas that seems to me that could improve the user experience, but the package looks great as it is. Nice testing suite!**

- In the README, the authors use an abbreviation (FAIR). It might be good to clarify what is about (Findability, Accessibility, Interoperability, and Reusability) for readers that are not introduced to  the concept of FAIR. Also, what do the authors think about linking a text to clarify it?

#### Vignette

- In the [vignette](https://frictionlessdata.github.io/frictionless-r/articles/frictionless.html),  the pipe from `magrittr` is used but it was not loaded. So if someone is trying to follow the code in the vignette, there is an `could not find function "%>%"` error.

- Just for the record: When executing `library(frictionless)` there is a message saying that `The following object is masked from ‘package:usethis’: create_package`: https://usethis.r-lib.org/reference/create_package.html

- There is some examples that are shown using lists. What do the authors think about using functions instead, to be more standardized? Example: create a function  `resources(package)` to use it instead of `package$resource_names`.

- Adding the descriptions to the schema does not seem trivial. There is an example with the `purrr` package. But the example might be not simple to understand if someone is not used to the `purrr` package.

I`m talking about this piece of code:

```r
iris_schema <- create_schema(iris)

# Remove description for first field
iris_schema$fields[[1]]$description <- NULL

# Set descriptions for all fields
descriptions <- c(
  "Sepal length in cm.",
  "Sepal width in cm.",
  "Pedal length in cm.",
  "Pedal width in cm.",
  "Iris species."
)
iris_schema$fields <- purrr::imap(
  iris_schema$fields,
  ~ c(.x, description = descriptions[.y])
)
```
Do the authors think that it is possible to create a function to add descriptions to the schema, in a way that is used in a similarly to  the other functions of the package? Example of the idea:

```r
iris_schema <- create_schema(iris) |>
  add_description(
    c(
      "Sepal length in cm.",
      "Sepal width in cm.",
      "Pedal length in cm.",
      "Pedal width in cm.",
      "Iris species."
    )
  )
```

- The vignette says that in order to validate the Data Package, the users need to use Python. Is there any plans to implement that funcionality in R aswell?

> Validate your Data Package before depositing. You can do this in Python with the Frictionless Framework using frictionless validate datapackage.json.

- In the vignette, there is some instructions to zip the csv to reduce size of the files:

> Zip the individual csv files (and update their paths in datapackage.json) to reduce size, not the entire Data Package. That way, users still have direct access to the datapackage.json file. See this example.

Do the authors think that is a good idea to add an argument on `write_package()` to write the compressed csvs, to facilitate this step? Something like:

```r
write_package(my_package, "my_directory", compress = TRUE)
```

[Since the function uses `readr::write_csv()`](https://github.com/frictionlessdata/frictionless-r/blob/a446512da1ff00384f06064d3cdfab1d3ce2d29f/R/write_resource.R#L18), if `compress = TRUE` it could add `.gz` at the end of the filepath and readr would zip it on the fly.

```r
readr::write_csv(mtcars, "mtcars.csv")
file.size("mtcars.csv")
#> [1] 1281

readr::write_csv(mtcars, "mtcars_zipped.csv.gz")
file.size("mtcars_zipped.csv.gz")
#> [1] 558
```

I hope the review is usefull. Again, congratulations for the authors.

Full report is here: https://github.com/beatrizmilz/ropensci_reviews/blob/main/frictionless/review.md
