### 2022-01-31 frictionlessdata/frictionless-r#97 @peterdesmet
Create function `resources(package)`
Suggested by @beatrizmilz in https://github.com/ropensci/software-review/issues/495#issuecomment-1025860861

> There is some examples that are shown using lists. What do the authors think about using functions instead, to be more standardized? Example: create a function `resources(package)` to use it instead of `package$resource_names`.
