### 2022-01-31 frictionlessdata/frictionless-r#98 @peterdesmet
Support writing compressed resources with `write_package()`
Suggested by @beatrizmilz in https://github.com/ropensci/software-review/issues/495#issuecomment-1025860861:

Do the authors think that is a good idea to add an argument on write_package() to write the compressed csvs, to facilitate this step? Something like:

```R
write_package(my_package, "my_directory", compress = TRUE)
```

[Since the function uses `readr::write_csv()`](https://github.com/frictionlessdata/frictionless-r/blob/a446512da1ff00384f06064d3cdfab1d3ce2d29f/R/write_resource.R#L18), if `compress = TRUE` it could add `.gz` at the end of the filepath and readr would zip it on the fly.

Example:

```R
readr::write_csv(mtcars, "mtcars.csv")
file.size("mtcars.csv")
#> [1] 1281

readr::write_csv(mtcars, "mtcars_zipped.csv.gz")
file.size("mtcars_zipped.csv.gz")
#> [1] 558
```
