### 2022-01-31 4a1da90
Add URL for term "FAIR"
