+++ README.md
@@ -19,7 +19,8 @@ Frictionless is an R package to read and write Frictionless Data
 Packages. A [Data
 Package](https://specs.frictionlessdata.io/data-package/) is a simple
 container format and standard to describe and package a collection of
-(tabular) data. It is typically used to publish FAIR and open datasets.
+(tabular) data. It is typically used to publish
+[FAIR](https://www.go-fair.org/fair-principles/) and open datasets.
 
 To get started, see:
 
