+++ man/frictionless-package.Rd
@@ -28,6 +28,7 @@ Authors:
 Other contributors:
 \itemize{
   \item Research Institute for Nature and Forest (INBO) \email{info@inbo.be} [copyright holder]
+  \item Beatriz Milz \email{milz.bea@gmail.com} (\href{https://orcid.org/0000-0002-3064-4486}{ORCID}) [reviewer]
 }
 
 }
