### 2022-01-31 f0d1b65
Support compress parameter
Fix #98
