+++ NEWS.md
@@ -3,6 +3,8 @@
 - `add_resource()` now sets `format`, `mediatype` and `encoding` for added CSV 
   file(s) (#78).
 - `add_resource()` now supports adding `schema` via path or URL.
+- `write_package()` now supports added data to be gzip compressed before being
+  written to disk (#98).
 - `read_resource()` will now warn rather than error on unknown encoding (#86).
 - `package` objects no longer have or require the custom attribute 
   `datapackage`, making it easier to edit them as lists (with e.g. `append()`).
