### 2022-01-31 frictionlessdata/frictionless-r#98 @peterdesmet
@damianooldoni this is now included as a PR in #99. At first I thought we should give the option to compress CSV files at resource level, rather than a package level, but decided against this:

- It (also) makes sense at a package level. Note that it would only be activated for _added_ resources (not existing ones that already have a corresponding csv file).
- Whether to zip or not is more of a decision at the "writing" then add the "adding" stage. E.g. do I write to disk so I can inspect the file or am I making a package to be deposited (where size matters)?
- `add_resource()` is already fairly complex, so it's good not to add another property there.
- Fine grained control over which resources to zip and which not to zip can be done by providing a path (rather than a data frame) to `add_resource()`.
