### 2022-01-31 frictionlessdata/frictionless-r#97 @peterdesmet
@damianooldoni especially for getting `package$resource_names` I like this suggestion, as is it something a user almost always has to do when exploring a package. I'm just wondering what name would be best:

```R
resources(example_package)
#> [1] "deployments"  "observations" "media"

resource_names(example_package)
#> [1] "deployments"  "observations" "media"
```
