### 2022-02-01 frictionlessdata/frictionless-r#97 @damianooldoni
Indeed, interesting suggestion by @beatrizmilz!

`list_resources()` and `resource_names` are too long for what they do and I tend to forget them fast.

I think easier the better in this case, so I vote for the first one.

I have also a even more simple option in mind: why not using the singular version of it, i.e. `resource()`? The reason is that all functions are singular: `add_resource`, `read_resource`, `remove_resource`.
