### 2022-02-01 frictionlessdata/frictionless-r#97 @peterdesmet
Thanks, I prefer `resources()` too. The singular functions always act on a single item, `resources()` returns a vector of (often multiple) resources, so I would keep it plural.

Another question is where to [group this function](https://frictionlessdata.github.io/frictionless-r/reference/index.html) (i.e. function family). I'm tempted to say this is a `read_` function, not a miscellaneous one:

I think I'll write the function title as "List Data Resources" not "List Data Resource names"
