### 2022-02-01 frictionlessdata/frictionless-r#97 @damianooldoni
Yes, Peter you are right: `resources` is the best name then.

And yes, it's definitely a read function, I agree 100%
