### 2022-02-01 0ab5cba
Create resources() function
