### 2022-02-01 73f9de8
Create resources() function
