+++ NAMESPACE
@@ -7,4 +7,5 @@ export(get_schema)
 export(read_package)
 export(read_resource)
 export(remove_resource)
+export(resources)
 export(write_package)
