### 2022-02-01 892af96
Use resources() throughout documentation
