+++ R/write_package.R
@@ -29,7 +29,7 @@
 #' )
 #'
 #' # List the resource names
-#' package$resource_names
+#' resources(package)
 #'
 #' # Write the (unchanged) Data Package to disk
 #' write_package(package, directory = "my_directory")
