+++ man/remove_resource.Rd
@@ -24,13 +24,13 @@ from a Data Package, i.e. it removes one of the described \code{resources}.
 package <- example_package
 
 # List the resource names
-package$resource_names
+resources(package)
 
 # Remove the resource "observations"
 package <- remove_resource(package, "observations")
 
 # List the resource names ("observations" removed)
-package$resource_names
+resources(package)
 }
 \seealso{
 Other edit functions: 
