### 2022-02-01 ef98406
Use resources() throughout documentation
