### 2022-02-01 3704015
Drop custom package$resource_names property
Now that we have resources() function, it's no longer necessary to store this information _in_ the Data Package list object (which could potentially confuse users). The only custom property left is "directory".

Removing it also means:
- create_package(), read_package() don't need to add it
- add_resource(), remove_resource() don't need to update it
- check_package() doesn't need to check if it is out of sync
- write_package() doesn't need to remove it before writing to disk
