+++ R/resources.R
@@ -1,9 +1,9 @@
 #' List Data Resources
 #'
-#' List the names of the Data Resources included in a Data Package
+#' Lists the names of the Data Resources included in a Data Package.
 #'
 #' @param package List describing a Data Package.
-#' @return Character vector with the resource names.
+#' @return Character vector with the Data Resource names.
 #' @family read functions
 #' @export
 #' @examples
@@ -13,6 +13,9 @@
 #' # List the resource names
 #' resources(package)
 resources <- function(package) {
+  # Check package (and that all resource have a name)
   check_package(package)
-  package$resource_names
+
+  # Get resource names
+  purrr::map_chr(package$resources, "name")
 }
