+++ R/check_package.R
@@ -16,14 +16,13 @@ check_package <- function(package) {
   # Check package is list with required properties
   assertthat::assert_that(
     is.list(package) &
-      all(c("resources", "resource_names", "directory") %in% names(package)),
+      all(c("resources", "directory") %in% names(package)),
     msg = msg_invalid
   )
 
   # Check package properties have correct class
   assertthat::assert_that(
     is.list(package$resources) &
-      is.character(package$resource_names) &
       is.character(package$directory),
     msg = msg_invalid
   )
@@ -33,20 +32,6 @@ check_package <- function(package) {
     purrr::every(package$resources, ~ !is.null(.x$name)),
     msg = glue::glue(
       "All resources in `package` must have property `name`."
-    )
-  )
-
-  # Check resource_names are in sync with resources name
-  unknown_names <- setdiff(
-    package$resource_names, purrr::map_chr(package$resources, ~ .x$name)
-  )
-  unknown_names_collapse <- paste(unknown_names, collapse = "`, `")
-  assertthat::assert_that(
-    length(unknown_names) == 0,
-    msg = glue::glue(
-      "Can't find resource(s) with name(s) `{unknown_names_collapse}`.",
-      "\u2139 Is `package$resource_names` out of sync with `name` of resources?",
-      .sep = "\n"
     )
   )
 }
