### 2022-02-01 00dc202
Rephrase in examples to "list resources"
