+++ man/read_resource.Rd
@@ -200,7 +200,7 @@ package <- read_package(
   system.file("extdata", "datapackage.json", package = "frictionless")
 )
 
-# List the resource names
+# List resources
 resources(package)
 
 # Read data from the resource "observations"
