### 2022-02-01 50f01b4
Rephrase in examples to "list resources"
