+++ R/write_package.R
@@ -25,7 +25,7 @@
 #'   system.file("extdata", "datapackage.json", package = "frictionless")
 #' )
 #'
-#' # List the resource names
+#' # List resources
 #' resources(package)
 #'
 #' # Write the (unchanged) Data Package to disk
