+++ R/read_resource.R
@@ -178,7 +178,7 @@
 #'   system.file("extdata", "datapackage.json", package = "frictionless")
 #' )
 #'
-#' # List the resource names
+#' # List resources
 #' resources(package)
 #'
 #' # Read data from the resource "observations"
