+++ man/add_resource.Rd
@@ -47,7 +47,7 @@ The resource name can only contain lowercase alphanumeric characters plus
 # Load the example Data Package
 package <- example_package
 
-# List the resource names
+# List resources
 resources(package)
 
 # Create a data frame
@@ -72,7 +72,7 @@ path_1 <- system.file("extdata", "observations_1.csv", package = "frictionless")
 path_2 <- system.file("extdata", "observations_2.csv", package = "frictionless")
 package <- add_resource(package, "observations_2", data = c(path_1, path_2))
 
-# List the resource names ("positions", "positions_2", "observations_2" added)
+# List resources ("positions", "positions_2", "observations_2" added)
 resources(package)
 }
 \seealso{
