+++ README.md
@@ -58,7 +58,7 @@ package <- read_package("https://zenodo.org/record/5879096/files/datapackage.jso
 #> Follow applicable norms or requirements to credit the dataset and its authors.
 #> For more information, see https://doi.org/10.5281/zenodo.5879096
 
-# List the available resources
+# List resources
 resources(package)
 #> [1] "reference-data" "gps"            "acceleration"
 
