+++ README.Rmd
@@ -53,7 +53,7 @@ library(frictionless)
 # reading them, which is convenient and fast.
 package <- read_package("https://zenodo.org/record/5879096/files/datapackage.json")
 
-# List the available resources
+# List resources
 resources(package)
 
 # Read data from the resource "gps"
