### 2022-02-01 74f16d6
Update NEWS.md
