### 2022-02-01 7d1e30f
Update NEWS.md
