+++ README.Rmd
@@ -83,7 +83,7 @@ For more functionality, see [get started](https://frictionlessdata.github.io/fri
 
 ## frictionless vs datapackage.r
 
-[datapackage.r](https://cran.r-project.org/web/packages/datapackage.r/) is an alternative R package to work with Data Packages. It has an object-oriented design (using a `Package` class) and offers validation. frictionless on the other hand allows you to quickly read and write Data Packages to and from data frames, getting out of the your for the rest of your analysis. It is designed to be lightweight, follows [tidyverse](https://www.tidyverse.org/) principles and supports piping.
+[datapackage.r](https://cran.r-project.org/web/packages/datapackage.r/) is an alternative R package to work with Data Packages. It has an object-oriented design (using a `Package` class) and offers validation. frictionless on the other hand allows you to quickly read and write Data Packages to and from data frames, getting out of the way for the rest of your analysis. It is designed to be lightweight, follows [tidyverse](https://www.tidyverse.org/) principles and supports piping.
 
 ## Meta
 
