### 2022-02-08 #495 @zambujo
Dear all, apologies for the delay. Please find my comments below:

Estimated hours spent reviewing: 4h

---

The package integrates the *frictionlessdata* framework library collection which facilitates the packaging of tabular text data along with their schemas across different programming languages.  The framework provides a set of tools intended to facilitate the creation of "FAIR-compliant" datasets. The umbrella project is led by the Open Knowledge Foundation and the framework is most known for its command-line tool written in Python.

Regarding the R package, the documentation is well organised and complete.  The same applies to the unit tests. (Apropos, I like seeing how the authors handle errors with abundant assertions directly on the main code.) I was unable to find any relevant issues and have only a few minor **optional** suggestions as well as some open points/questions for discussion.  All in all, the package has been beautifully crafted.  Well done!

#### Optional suggestions/questions:

(in no particular order)

- `check_path(path)` (in `utils.R`): would it make sense to include other protocols other than http by using something like `grepl("^http://|^https://|^ftp://|sftp://", path)` - instead of `starts_with("http")`
- in `unique_sorted()` (in `utils.R`), I was wondering whether `stats::aggregate()` is necessary.  Intuitively I would have used `table()` from base R: `names(sort(table(x), decreasing = TRUE))`, provided it passes the unit test.
- Is there a reason to pass a default value to the file name parameter in `read_package()`? Certain functions such as `bmp()` provide names by default with automatic file numbering, but I think it is more common not to provide any default value.
- would it make sense to replace `{readr}` with `{vroom}` to improve reading and writing times for large files?
- if I understand correctly, when creating a data package, factors are converted to strings of characters. Would it make sense to package extra dedicated lookup tables to account for the information contained factor levels and their order?
- in the future, would it make sense to develop `{frictionless-r}` more towards being a `frictionless-py` wrapper (somewhat similarly to [`spacyr`](https://github.com/quanteda/spacyr)/[`spacy`](https://spacy.io)), to make it easier to keep the R package in sync with the Python project?
