### 2022-02-09 fe2d550
Use longer example in unique_sorted doc
