+++ R/utils.R
@@ -17,8 +17,9 @@ replace_null <- function(x, replacement) {
 
 #' Get unique vector values sorted by how often they occur
 #'
-#' @param x Vector, e.g. `c("a", "b", "b", "b", "a")`.
-#' @return Vector with unique values sorted by desc. count, e.g. `c("b", "a")`.
+#' @param x Vector, e.g. `c("a", "b", "b", "b", "c", "a")`.
+#' @return Vector with unique values sorted by most to least occurring,
+#'   e.g. `c("b", "a", "c")`.
 #' @family helper functions
 #' @noRd
 unique_sorted <- function(x) {
