### 2022-02-09 #495 @peterdesmet
Thanks @zambujo for your review. My feedback:

- **Other protocols for `check_path()`**: Other protocols like FTP could indeed be implemented, but the [specs](https://specs.frictionlessdata.io/data-resource/#url-or-path) for "URL or path" states that only http, https, and local POSIX paths are allowed. It makes sense though to allow (S)FTP, so have asked the Frictionless community for guidance.
- **Using `table()` in `unique_sorted()`**: Nice! More elegant indeed. Only had to update to handle all `NA_character_` values.
- **No default value for `read_package()`**: Although it is unlikely that a `datapackage.json` will be present in the working directory, I'm tempted to keep it, because according to the [specs](https://specs.frictionlessdata.io/data-package/#descriptor) a descriptor file must be named `datapackage.json`, so the default name hints that the user should provide a path to such a file.
- **vroom**: `{readr}` 2.0.0 and up use `{vroom}` under the hood. `frictionless-r` requires `readr >= 2.1.0` and thus `{vroom}`, so I don't think there is going to be a speed difference using `{vroom}` directly. Even though many functions in `{readr}` can be exchanged for `{vroom}` functions, I'll keep `{readr}` because 1) I communicate in the documentation that `{readr}` is used by some functions and that is a more well-known package to end users than `{vroom}` and 2) I rely on `readr::guess_encoding()`.
- **factors**: when _reading_ a data package, str/integer/numeric values that have an `enum` in their schema are converted to factors, with the `enum` values as `levels`,  in the order they are listed in `enum`. When _writing_ a data package, factors keep their data type and the levels are written in an `enum` field. No (re)ordering is done.
    ```json
    {
      "name": "str_factor",
      "type": "string",
      "constraints": {
        "enum": ["foo", "bar"]
      }
    },
    {
      "name": "num_factor",
      "type": "number",
      "constraints": {
        "enum": [3.1, 3.2, 3.3]
      }
    },
    {
      "name": "int_factor",
      "type": "integer",
      "constraints": {
        "enum": [3, 4, -1]
      }
    }
   ```
- **r package as wrapper around Python**: An option indeed, but:
  1. Currently out of my area of expertise
  2. Python dev is going fast and currently a bit of a moving target
  3. The main target to keep in sync with are the specs and they luckily don't change that fast.
  4. In my opinion the previous R package `{datapackage.r}` suffers from being un-R-like, introducing OO concepts that are foreign to most R users, which might happen again on a wrapper.
However, for Data Package _validation_ specifically (currently not in scope for `{frictionless-r}`), wrapping around the Python toolbox is likely useful!
