### 2022-02-09 #495 @zambujo
Many thanks @peterdesmet. You have addressed all my comments and questions. Impressive work!

> Ps. I have to confess that I had to update my packages to be able to review `frictionless-r`. Incidentally, I did notice a huge improvement in the performance of `{readr}` when I ran some code this morning. 🤓
