### 2022-02-09 f9dbf1d
Use table() in unique_sorted() and add more tests
