### 2022-02-09 #495 @peterdesmet
Thanks @zambujo. The suggested change for `unique_values()` is implemented in https://github.com/frictionlessdata/frictionless-r/pull/101.

@melvidoni, the comments suggested by @beatrizmilz are addressed in https://github.com/ropensci/software-review/issues/495#issuecomment-1025957514 and where actionable, all implemented in the latest version of the package. Both reviewers were included with `rev` roles: thanks to you both!

One lingering question I have for the reviewers is the use of the word `package`. I'm copy/pasting my question from higher up:

> - **Masking of `usethis::create_package()`**: Yeah, it is a bit unfortunate that the term `package` is used for different things in Frictionless vs R (as explained at the start of the vignette). Luckily in R it is often referred to as `pkg` in function names, reducing masking. In the Frictionless Community "Data Package" does seem to be consistently referred to as `package` in implementations in other languages, not `dp`, seldom as `datapackage`, which is why I adopted that term for `frictionless` functions and parameters. I think alternatives like `create_datapackage()`, `create_data_package()`, `create_dataset()` are less desirable, but 👉 feedback welcome 👈. Was the term `package` confusing in any way?

Since you both didn't remark on that, I assume that the word `package` was not confusing in `read_package()`, `create_package()` or `write_package()`, but I want to make sure.
