### 2022-02-09 20c91cf
Include pkg review suggestions by @zambujo (#101)
* Use longer example in unique_sorted doc
* Use table() in unique_sorted() and add more tests
* Add @zambujo as reviewer
