### 2022-02-10 #495 @beatrizmilz
Hi! Peter, the word `package` was not confusing for me since there was an explaination in the documentation! I  pointed out about the masking with `usethis` because for users that uses only `library()` and are not familiar with the possibility of conflict between functions with the same names, can eventually encounter errors and ask for help (and that is not a problem with the package!).  I think i thought about that because I answer a lot of questions in foruns in portuguese, and is frequent questions about errors caused by masking.

You have addressed all the questions, and as @zambujo , this is an impressive work. Congratulations!
