### 2022-02-10 #495 @ropensci-review-bot
Approved! Thanks @peterdesmet for submitting and @zambujo, @beatrizmilz for your reviews! :grin:
