### 2022-02-10 #495 @peterdesmet
@melvidoni Is it required to transfer the `{frictionless-r}` repository to rOpenSci? **Because this is also a question of branding**. Ping @lwinfree @sapetti9 @roll from Frictionless Data.

- [ ] **Package repository URL**: I would prefer to keep https://github.com/frictionlessdata/frictionless-r .That way 1) it is clearer that it is endorsed/maintained by Frictionless Data and 2) it keeps living alongside other implementations of Frictionless Data standards, such [Python](https://github.com/frictionlessdata/frictionless-py) and [JS](https://github.com/frictionlessdata/frictionless-js).
- [ ] **Package website**: currently https://frictionlessdata.github.io/frictionless-r/ with generic pkgdown (Bootstrap v5). We could adopt the rOpenSci central docs building and branding, so it looks more like https://docs.ropensci.org/wateRinfo/. That will visually tie it to rOpenSci and required an update of its URL (including a redirect of the old URL). Branding wise, my opinion is that it is fine to adopt the rOpenSci branding, because I hope to add a logo to the package soon, which will visually tie it to Frictionless Data.
- [ ] **rOpenSci package family**: how can I get the package listed under https://ropensci.org/packages/all/  ?
- [ ] **rOpenSci R universe**: how can I get the package added to the rOpenSci R universe?
- [ ] **Code of conduct**: Which code of conduct should we adopt? [Frictionless Data](https://frictionlessdata.io/work-with-us/code-of-conduct) vs [rOpenSci](https://ropensci.org/code-of-conduct/) I think Frictionless Data, since the package is maintained there and conflicts should be resolved there.

Once those questions are answered I can make the necessary changes and then hopefully submit to CRAN! 🎉🤞
