### 2022-02-10 #495 @lwinfree
Hi all! First of all, it has been really lovely to watch this process unfold, so a big thank you to everyone that has been involved!

Speaking as product manager of Frictionless Data:
- we would like to keep the repo under the Frictionless organization as Peter suggests
- we would like to keep the Frictionless code of conduct to not confuse potential users if that is OK

everything else looks great to me!

Thanks!
