### 2022-02-15 #495 @maelle
* https://docs.ropensci.org/frictionless/ is now up. It means the package has been added to rOpenSci R-universe https://ropensci.r-universe.dev/ui#builds For reference, here's how it was "registered": https://github.com/ropensci-org/makeregistry/commit/31e1638913977f508ef0330879b8cadb45c58f13
* it is fine it links to the frictionless COC. You might want to add that COC to https://github.com/frictionlessdata/frictionless-r/tree/main/.github to have it appear in the sidebar (or customize the sidebar by other means see https://pkgdown.r-lib.org/reference/build_home.html#sidebar).
* the package will soon also be on our website, whose registry is built once a day.
* Regarding redirects we have some guidance at https://devguide.ropensci.org/redirect.html (and if it's unclear please tell us so that we might clarify it!)

Thanks for your work on this package. :smile_cat:
