### 2022-02-15 #495 @maelle
![screenshot from rOpenSci website where one gets a card for the frictionless package](https://user-images.githubusercontent.com/8360597/154037181-153c307c-c813-46da-814e-f36d11649345.png)

https://ropensci.org/packages/all/
