### 2022-02-15 #495 @peterdesmet
Thanks @maelle!

- I've added a small `.github/CODE_OF_CONDUCT.md` page that points to Frictionless Data COC. This makes it appear in sidebar.
- Rather than setting up a `frictionlessdata.github.io` repository (required when the repository moves organizations - not the case here), I have placed a redirect `index.html` on the `gh-pages` branch of the repository, disabled automatic pkgdown building, but kept GitHub Pages active. https://frictionlessdata.github.io/frictionless-r/ now successfully redirects
- I have kept some [minor styling tweaks](https://github.com/frictionlessdata/frictionless-r/blob/main/_pkgdown.yml#L3-L7) to `_pkgdown.yml` for local pkgdown building. I don't think it will affect building at rOpenSci docs.
- I have kept the other GitHub Actions (e.g. `test-coverage.yaml`)  [intact](https://github.com/frictionlessdata/frictionless-r/tree/main/.github/workflows). **Or are those checks provided by rOpenSci CI and thus not necessary in the repository**?
