### 2022-02-15 #495 @maelle
Thank you!

> I have kept some [minor styling tweaks](https://github.com/frictionlessdata/frictionless-r/blob/main/_pkgdown.yml?rgh-link-date=2022-02-15T11%3A39%3A28Z#L3-L7) to _pkgdown.yml for local pkgdown building. I don't think it will affect building at rOpenSci docs.

Indeed, we override those in https://github.com/ropensci-org/rotemplate

> I have kept the other GitHub Actions (e.g. test-coverage.yaml) [intact](https://github.com/frictionlessdata/frictionless-r/tree/main/.github/workflows?rgh-link-date=2022-02-15T11%3A39%3A28Z). Or are those checks provided by rOpenSci CI and thus not necessary in the repository?

It is good to keep them indeed. R-universe does build the package but you wouldn't get notified and you can't share credentials for instance. You can see the R-universe status of your package at https://ropensci.r-universe.dev/ui#builds
